# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/loan-offerable-symbols/{symbol-code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get symbol not active blacklist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "symbol",
                    "admin"
                ],
                "summary": "Get symbol not active blacklist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "symbol-code",
                        "name": "symbol-code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_Symbol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/my-symbols/{symbol}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get symbol by symbol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "symbol",
                    "investor"
                ],
                "summary": "Get symbol by symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "symbol",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_Symbol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/symbols": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all symbols",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "symbol",
                    "admin"
                ],
                "summary": "Get all symbols",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ResponseWithPaging-array_financing-offer_internal_core_entity_Symbol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create symbol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "symbol",
                    "admin"
                ],
                "summary": "Create symbol",
                "parameters": [
                    {
                        "description": "symbol",
                        "name": "symbol",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_core_symbol_transport_http.SymbolRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_Symbol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/symbols/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get symbol by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "symbol",
                    "admin"
                ],
                "summary": "Get symbol by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_Symbol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update symbol status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "symbol",
                    "admin"
                ],
                "summary": "Update symbol",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_core_symbol_transport_http.UpdateSymbolStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_Symbol"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "financing-offer_internal_core.PagingMetaData": {
            "type": "object",
            "properties": {
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "financing-offer_internal_core_entity.AssetType": {
            "type": "string",
            "enum": [
                "UNDERLYING",
                "DERIVATIVE"
            ],
            "x-enum-varnames": [
                "AssetTypeUnderlying",
                "AssetTypeDerivative"
            ]
        },
...
        "financing-offer_internal_core_entity.Symbol": {
            "type": "object",
            "properties": {
                "assetType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastUpdatedBy": {
                    "type": "string"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScore"
                    }
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolStatus"
                },
                "stockExchangeId": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.SymbolScore": {
            "type": "object",
            "properties": {
                "affectedFrom": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "INACTIVE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScoreStatus"
                        }
                    ]
                },
                "symbolId": {
                    "type": "integer"
                },
                "type": {
                    "enum": [
                        "MANUAL",
                        "SYSTEM"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScoreType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.SymbolScoreStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "SymbolScoreStatusActive",
                "SymbolScoreStatusInactive"
            ]
        },
        "financing-offer_internal_core_entity.SymbolScoreType": {
            "type": "string",
            "enum": [
                "MANUAL",
                "SYSTEM"
            ],
            "x-enum-varnames": [
                "SymbolScoreTypeManual",
                "SymbolScoreTypeSystem"
            ]
        },
        "financing-offer_internal_core_entity.SymbolStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "SymbolStatusActive",
                "SymbolStatusInactive"
            ]
        },
...
        "financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_Symbol": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.Symbol"
                }
            }
        },
...
        "financing-offer_internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "financing-offer_internal_handler.ResponseWithPaging-array_financing-offer_internal_core_entity_Symbol": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.Symbol"
                    }
                },
                "metaData": {
                    "$ref": "#/definitions/financing-offer_internal_core.PagingMetaData"
                }
            }
        },
...
        "internal_core_symbol_transport_http.SymbolRequest": {
            "type": "object",
            "required": [
                "assetType",
                "stockExchangeId",
                "symbol"
            ],
            "properties": {
                "assetType": {
                    "enum": [
                        "UNDERLYING",
                        "DERIVATIVE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                        }
                    ]
                },
                "stockExchangeId": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "internal_core_symbol_transport_http.UpdateSymbolStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "enum": [
                        "ACTIVE",
                        "INACTIVE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolStatus"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}